% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoparser.R
\name{geoparser}
\alias{geoparser}
\title{Detect countries names}
\usage{
geoparser(x)
}
\arguments{
\item{x}{a \code{character} in which countries will be detected.}
}
\value{
a \code{data.frame} with the following four columns:
\itemize{
\item \code{geographic_entity}: the name of the country
\item \code{n_pages}: the total number of pages in the document
\item \code{page}: the page number
\item \code{count}: the occurrence of the country for a given page
}
}
\description{
Detect countries names
}
\examples{
## ----
}
