% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geoparser_key}
\alias{geoparser_key}
\title{Retrieve Geoparser.io API key}
\usage{
geoparser_key(quiet = TRUE)
}
\description{
A Geoparser.io API Key
Looks in env var \code{GEOPARSER_KEY}
}
\keyword{internal}

