% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoparser_query.R
\name{geoparser_q}
\alias{geoparser_q}
\title{Geoparser query}
\usage{
geoparser_q(text_input, key = geoparser_key())
}
\arguments{
\item{text_input}{A string, whose size must be smaller than 8 KB. See \code{nchar(text_input, type = "bytes")}.}

\item{key}{Your Geoparser.io key}
}
\value{
A data.frame (dplyr tbl_df)
}
\description{
Geoparser query
}
\details{
To get an API key, you need to register at \url{https://geoparser.io/pricing.html}.
With an hobbyist account, you can make up to 1,000 calls a month to the API.
}
\examples{
geoparser_q(text_input = "I was born in Vannes but I live in Barcelona.")
}

