% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoparser.R
\name{geoparser}
\alias{geoparser}
\title{Detect countries names}
\usage{
geoparser(x)
}
\arguments{
\item{x}{a \code{character} in which countries will be detected.}
}
\value{
a \code{data.frame} with: ...
}
\description{
Detect countries names
}
\examples{
## ----
}
